Made image2vectorCNN class, moved generalMLP class to cnn_utils
