+++ src/yoke/models/cnn_utils.py
@@ -2,6 +2,11 @@
 
 import math
 
+from collections import OrderedDict
+
+import torch
+import torch.nn as nn
+
 
 ####################################
 # Get Conv2D Shape
@@ -86,6 +91,83 @@ def convtranspose2d_shape(
     total = new_w * new_h
 
     return new_w, new_h, total
+
+
+class generalMLP(nn.Module):
+    """A general multi-layer perceptron structure.
+
+    Consists of stacked linear layers, normalizing layers, and
+    activations. This is meant to be reused as a highly customizeable, but
+    standardized, MLP structure.
+
+    Args:
+        input_dim (int): Dimension of input
+        output_dim (int): Dimension of output
+        hidden_feature_list (tuple[int, ...]): List of number of features in each layer.
+                                               Length determines number of layers.
+        act_layer (nn.modules.activation): torch neural network layer class to
+                                           use as activation
+        norm_layer (nn.Module): Normalization layer.
+
+    """
+
+    def __init__(
+        self,
+        input_dim: int = 64,
+        output_dim: int = 16,
+        hidden_feature_list: tuple[int, ...] = (16, 32, 32, 16),
+        act_layer: nn.Module = nn.GELU,
+        norm_layer: nn.Module = nn.LayerNorm,
+    ) -> None:
+        """Initialization for MLP."""
+        super().__init__()
+
+        self.input_dim = input_dim
+        self.output_dim = output_dim
+        self.hidden_feature_list = hidden_feature_list
+        self.act_layer = act_layer
+        self.norm_layer = norm_layer
+
+        # Create full feature list without mutating input
+        self.feature_list = (input_dim,) + hidden_feature_list + (output_dim,)
+
+        # Module list to hold linear, normalization, and activation layers.
+        self.LayerList = nn.ModuleList()
+        # Create transpose convolutional layer for each entry in feature list.
+        for i in range(len(self.feature_list) - 1):
+            linear = nn.Linear(self.feature_list[i], self.feature_list[i + 1])
+
+            normalize = self.norm_layer(self.feature_list[i + 1])
+            activation = self.act_layer()
+
+            # Make list of small sequential modules. Then we'll use enumerate
+            # in forward method.
+            #
+            # Don't attach an activation to the final layer
+            if i == len(self.feature_list) - 2:
+                cmpd_dict = OrderedDict(
+                    [
+                        (f"linear{i:02d}", linear),
+                    ]
+                )
+            else:
+                cmpd_dict = OrderedDict(
+                    [
+                        (f"linear{i:02d}", linear),
+                        (f"norm{i:02d}", normalize),
+                        (f"act{i}", activation),
+                    ]
+                )
+
+            self.LayerList.append(nn.Sequential(cmpd_dict))
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        """Forward method for MLP."""
+        # enumeration of nn.moduleList is supported under `torch.jit.script`
+        for i, ll_layer in enumerate(self.LayerList):
+            x = ll_layer(x)
+
+        return x
 
 
 if __name__ == "__main__":
